feat: added information on recurrence exception dates in event subtitle (#829)
* feat: added information on recurrence exception dates in event subtitle

* fix: show condition

* fix: additional dates string
