+++ RELEASE.md
@@ -45,7 +45,7 @@
 
 ### Migliorie
 
-- ...
+- Se vengono aggiunte o rimosse date dalla ricorrenza di un CT evento, nel sottotitolo viene aggiunta l'indicazione di controllare le eccezioni nella sezione apposita.
 
 ### Novità
