+++ locales/de/LC_MESSAGES/volto.po
@@ -2070,6 +2070,11 @@ msgstr ""
 msgid "exceeded_chars"
 msgstr ""
 
+#: components/ItaliaTheme/View/Commons/PageHeader/PageHeaderEventDates
+# defaultMessage: con alcune eccezioni. Per maggiori informazioni controllare la sezione dedicata.
+msgid "exceptionDates"
+msgstr ""
+
 #: components/ItaliaTheme/View/Commons/PageHeader/PageHeaderDates
 # defaultMessage: Scadenza
 msgid "expire"
