+++ src/helpers/index.js
@@ -24,6 +24,7 @@ export {
   viewDate,
   getRealStartAndEndWithRecurrence,
   getRealEventEnd,
+  getRecurrenceExceptionDates,
 } from 'design-comuni-plone-theme/helpers/dates';
 export { getSiteProperty } from 'design-comuni-plone-theme/helpers/config';
 export { useDebouncedEffect } from 'design-comuni-plone-theme/helpers/debounce';
